Properly validate all metadata fields
This involved some hacks to get model creation forms to cooperate.
Django forms do not deal with foreign keys, which is a problem for
validation because models need to reference their dataset to get the
pattern used for validation. The solution is to pre-populate the form
with a model that already has all foreign keys set, and to highjack the
exclusion list passed to `validate_unique` to explicitly include the
foreign keys.