Add the vue app to the deployment.
This is a somewhat hacky way to add the vue app to the django deployment
at /static/index.html.  Ideally, we want to adjust it to serve as the
root (you'd still be able to access the admin pages) and handle
templates within the vue app.

This also temporarily disables typescript checking until an error with
the router file is resolved.